implement authorization at product page actions

+++ resources/views/product/list.blade.php
@@ -6,7 +6,9 @@
 <div class="card shadow">
   <div class="card-header py-3 d-flex flex-row align-items-center justify-content-between">
     <h6 class="m-0 font-weight-bold text-primary">Products List</h6>
-    <a href="/product/form" class="btn btn-primary btn-sm"><i class="fas fa-plus"></i></a>
+    @if (permission(auth()->user(),'product','add'))
+      <a href="/product/form" class="btn btn-primary btn-sm"><i class="fas fa-plus"></i></a>
+    @endif
   </div>
   <div class="card-body">
     <form action="" method="get" class="row">
@@ -52,8 +54,12 @@
               <td>{{ number_format($row->purchase_price) }}</td>
               <td>{!! ($row->status) ? '<span class="badge badge-success">Active</span>' : '<span class="badge badge-danger">Non Active</span>' !!}</td>
               <td width="100">
-                <a href="/product/form/{{ $row->id }}" class="btn btn-warning btn-sm btn-icon"><i class="fas fa-pen"></i></a>
-                <a href="/product/delete?id={{ $row->id }}" onclick="return confirm('Are you sure want to delete this product?')" class="btn btn-danger btn-sm btn-icon"><i class="fas fa-trash"></i></a>
+                @if (permission(auth()->user(),'product','edit'))
+                  <a href="/product/form/{{ $row->id }}" class="btn btn-warning btn-sm btn-icon"><i class="fas fa-pen"></i></a>
+                @endif
+                @if (permission(auth()->user(),'product','delete'))
+                  <a href="/product/delete?id={{ $row->id }}" onclick="return confirm('Are you sure want to delete this product?')" class="btn btn-danger btn-sm btn-icon"><i class="fas fa-trash"></i></a>
+                @endif
               </td>
             </tr>
           @endforeach
